+++ src/ACEtk/electron/BremsstrahlungBlock/test/BremsstrahlungBlock.test.cpp
@@ -1,7 +1,7 @@
 #define CATCH_CONFIG_MAIN
 
 #include "catch.hpp"
-#include "ACEtk/block/PhotoatomicElectronBremsstrahlungBlock.hpp"
+#include "ACEtk/electron/BremsstrahlungBlock.hpp"
 
 // other includes
