+++ src/ACEtk/electron/SubshellTransitionData/src/calculateNumberTransitions.hpp
